Update test runner for osx (#529)
* Update test runner for osx

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* lint

* fix up tests, simplify manager creation
