+++ tests/integration/test_random_beta.py
@@ -12,18 +12,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+import sys
+
 import numpy as np
 import pytest
 from utils.random import ModuleGenerator, assert_distribution
 
 import cunumeric as num
 
-BITGENERATOR_ARGS = [
-    ModuleGenerator,
-    num.random.XORWOW,
-    num.random.MRG32k3a,
-    num.random.PHILOX4_32_10,
-]
+if sys.platform == "darwin":
+    pytestmark = pytest.mark.skip()
+    BITGENERATOR_ARGS = []
+else:
+    BITGENERATOR_ARGS = [
+        ModuleGenerator,
+        num.random.XORWOW,
+        num.random.MRG32k3a,
+        num.random.PHILOX4_32_10,
+    ]
 
 
 @pytest.mark.parametrize("t", BITGENERATOR_ARGS, ids=str)
