+++ tests/integration/test_random_straightforward.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 #
 import math
+import sys
 
 import numpy as np
 import pytest
@@ -20,12 +21,16 @@ from utils.random import ModuleGenerator, assert_distribution
 
 import cunumeric as num
 
-BITGENERATOR_ARGS = [
-    ModuleGenerator,
-    num.random.XORWOW,
-    num.random.MRG32k3a,
-    num.random.PHILOX4_32_10,
-]
+if sys.platform == "darwin":
+    pytestmark = pytest.mark.skip()
+    BITGENERATOR_ARGS = []
+else:
+    BITGENERATOR_ARGS = [
+        ModuleGenerator,
+        num.random.XORWOW,
+        num.random.MRG32k3a,
+        num.random.PHILOX4_32_10,
+    ]
 
 
 @pytest.mark.parametrize("t", BITGENERATOR_ARGS, ids=str)
